Title: 특정 문자로 시작하는 문자열, Time: 141ms, Memory: 10MB, Status: Passed - CodeTree

+++ 240607/특정 문자로 시작하는 문자열/strings-that-start-with-a-specific-character.java
@@ -17,7 +17,7 @@ public class Main {
                 len+=arr[i].length();
             }
         }
-        float average=len/sum;
+        float average=(float)len/sum;
         System.out.print(sum+" ");
         System.out.printf("%.2f\n",average);
     }
